+++ miniprogram/pages/webview/webview.js
@@ -1,0 +1,7 @@
+Page({
+  data: { url: '' },
+  onLoad(query) {
+    const url = (query && query.url) ? decodeURIComponent(query.url) : '';
+    this.setData({ url });
+  },
+});